feat: allow configuration through workspace/didChangeConfiguration (#316)
## What kind of change does this PR introduce?

It enables passing workspace settings via the
workspace/didChangeConfiguration notification. This in turn enables the
client to specify settings dynamically, rather than being limited to
configuration files. This is _a_ solution to #302. See example usage
(with lspconfig) here:
https://github.com/willruggiano/neovim.drv/commit/9aa06ad7c889ed7b14d35ab34264e757d5ad1a7e.

## What is the current behavior?

There is none. The payload of this handler is currently ignored.

## What is the new behavior?

The configuration received by the handler is merged with the fs
configuration.

---------

Co-authored-by: Will Ruggiano <wmruggiano@gmail.com>