feat: versioned docs (#414)
(Finally) sets up versioned docs with a version selector. 

we deploy on every push to `main` on the `main` tag, and on every
release on the tag of that release + update `latest`

this finally fixes the issue we were having around missing schema files.

tested it on a fork. :) 

Todo
- [x] update schema url in default config 
- [x] Before merging: delete gh-pages branch 

closes #312
fixes #319