feat: plpgsql check (#469)
adds support for plpgsql_check. the approach is even simpler as
described in the issue: if we encounter a create function statement, we
start a transaction, run the statement, then run plpgsql_check on it,
and then rollback the transaction.

the remaining code is just about translating the location information we
get from the extension into a good range.

todo:
- [x] integrate into workspace api
- [x] to get span, move from the first word / occurrence in the line to
the next semicolon
- [x] handle return trigger by checking for all variations
- [x] make sure we do not report create fn syntax errors if
plpgsql_check is enabled
- [x] check why "not a known variable" is not getting the right span

closes #190