fix: positional params (#473)
pre-processes the query string to replace named params like `@name` with
positional params like `$1`, because only the latter is supported by the
postgres parser.

not super happy with the way it is used for the type checker but also
did not want to put that function in another crate.

also removed a `println!` statement leftover from a previous pr of me.

also fixed a bug in the lexer that tokenized a cast as a named
parameter.

fixes #405, #454