chore: offline builds with submodules (#486)
right now, we are cloning libpg_query on every build if its not
available locally. this is not only unnecessary, it also makes
reproducible builds even harder. this reverts that back to using
submodules instead.

also unified the directory structures across crates and removed the code
that was intended to support multi-version builds. I think its
unnecessary to even support that in our case since postgres syntax is
backward compatible, and `EXPLAIN` exists since 6.4.