fix: context-aware named param conversion (#513)
we were blindly converting named parameters to positional parameters.
but the latter is only valid as a literal, and not as an identifier.

statements like

```sql
grant usage on schema public, app_public, app_hidden to :DB_ROLE;
```

are not valid when converted to

```sql
grant usage on schema public, app_public, app_hidden to $1;
```

i went a bit back and forth on this and decided the easiest way to fix
this is to convert to identifiers like `a` if the previous non-trivia
token is one of a set list of tokens. We will probably have a bunch of
edge cases here but fixing them should be as easy as adding a keyword to
the list.

now, we convert to

```sql
grant usage on schema public, app_public, app_hidden to a;
```

closes #510