chore(pgt_query): support system build (#576)
after reviewing the comments on our first attempt to add this project to
homebrew (https://github.com/Homebrew/homebrew-core/pull/218049), i
realised that we can significantly improve the build in our homebrew
recipe if we add `libpg_query` as a homebrew dependency (yes, it is
available there as a formula).

this PR adds a "dual build" to the build script of `pgt_query`. If
`LIBPG_QUERY_PATH` is set, we link the library directly and skip the
build.

tested this locally.

once this is merged, I will publish a new release and create the pr for
homebrew.