feat(config): add connectionString (#573)
adds `connectionString` to the database config. if provided, it is
preferred over the individual `host`, `port`, etc. settings.

note that the `ConnectionKey` still checks only based on the basic
infos. This could potentially lead to an issue if the same workspace
instance accepts multiple connections to the same database. if one is
e.g. adding different `sslmode` settings via the connection sring, it
would be re-used even if another just has the same `host` + `port` +
`user` + `database`. I do not think we need to distinguish here since
that case should not really happen unless you host this as a service
that is shared among many people.

fixes #575